requirements.txt
requests==2.28.2
beautifulsoup4==4.12.0
discord.py==2.2.2
discord-ext-bot==1.0.1